lots of changes for moving and updating story
I needed to be able to apply special events before boarding a train, and
I needed the special events to apply before disembariking the train.
This has a lot of changes to make all of that work better, specifically
breaking up the update pipeline to include applyPendingChanges and
specialEvents.  I also handle Continue better.

Also has a few minor fixes that were bugs before that never occured.

As of now, the movemnt logic for the intro works.